vector-search-benchmarks/requirements.txt
numpy
faiss-cpu
psycopg2-binary
pgvector
